split up index.js
