+++ routes/helper_functions.js
@@ -1,0 +1,44 @@
+const fs = require("./imports").fs;
+
+let return_time = function () {
+    let date_ob = new Date();
+    let date = ("0" + date_ob.getDate()).slice(-2);
+    let month = ("0" + (date_ob.getMonth() + 1)).slice(-2);
+    let year = date_ob.getFullYear();
+    let hours = date_ob.getHours();
+    let minutes = date_ob.getMinutes();
+    let seconds = date_ob.getSeconds();
+    let dateDisplay = `${date}/${month}/${year} ${hours}:${minutes}:${seconds}`;
+    return dateDisplay;
+}
+
+let connection_log = function (text = "Connection from:", ip = '', time = '') {
+    let file_name = "connection_log.txt";
+    fs.appendFile(file_name, `${time} : ${text} : ${ip}\n`, () => {
+        console.log(`Saved to ${file_name}`);
+    });
+}
+
+let randomString = function (len) {
+    var buf = [],
+        chars = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789',
+        charlen = chars.length;
+
+    function getRandomInt(min, max) {
+        min = Math.ceil(min);
+        max = Math.floor(max);
+        return Math.floor(Math.random() * (max - min + 1)) + min;
+    }
+
+    for (var i = 0; i < len; ++i) {
+        buf.push(chars[getRandomInt(0, charlen - 1)]);
+    }
+
+    return buf.join('');
+}
+
+module.exports = {
+    return_time: return_time,
+    connection_log: connection_log,
+    randomString: randomString
+};
