+++ views/change_password.ejs
@@ -11,7 +11,7 @@
 <body>
     <%- include('navbar'); -%>
     <br>
-    <form action="/update_password" method="POST" class="login-form text-center">
+    <form action="/account/update_password" method="POST" class="login-form text-center">
         <input type="password" name="old_password" placeholder="Old Password" autocomplete="off" required>
         <input type="password" name="new_password" placeholder="New Password" autocomplete="off" required>
         <input type="password" name="confirm_new_password" placeholder="Confirm Password" autocomplete="off" required>
