+++ views/delete_account.ejs
@@ -11,7 +11,7 @@
 <body>
     <%- include('navbar'); -%>
     <br>
-    <form action="/delete" method="POST" class="login-form text-center">
+    <form action="/account/delete" method="POST" class="login-form text-center">
         <input type="password" name="password" placeholder="Enter Password" autocomplete="off" required>
         <button type="submit" class="text-center">Delete account</button>
     </form>
