+++ views/header.ejs
@@ -2,7 +2,6 @@
     <link rel="icon" href="https://omniamorsaequat.com/favicon.ico?id=31">
     <LINK REL="shortcut icon" HREF="https://omniamorsaequat.com/favicon.ico?id=69" TYPE="image/x-icon">
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" charset="UTF-8">
-    <meta name="google-site-verification" content="k9ABJOto26HFZwte5uGQ21IY_0FzSEnBpdlE_Wqu-RI" />
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css"
         integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO" crossorigin="anonymous">
 </head>
