+++ views/high_scores.ejs
@@ -13,16 +13,16 @@
 	<%- include('navbar'); -%>
 	<div class="container-fluid">
 		<br>
-		<a href="/high_scores/rocket_dodge">Rocket Dodge High Scores</a>
+		<a href="/score/high_scores/rocket_dodge">Rocket Dodge High Scores</a>
 		<br>
 		<br>
-		<a href="/high_scores/hold_dodge">Hold Dodge High Scores</a>
+		<a href="/score/high_scores/hold_dodge">Hold Dodge High Scores</a>
 		<br>
 		<br>
-		<a href="/high_scores/hold_dodge_accelerated">Hold Dodge (Accelerated) High Scores</a>
+		<a href="/score/high_scores/hold_dodge_accelerated">Hold Dodge (Accelerated) High Scores</a>
 		<br>
 		<br>
-		<a href="/high_scores/touch_dodge">Touch Dodge High Scores</a>
+		<a href="/score/high_scores/touch_dodge">Touch Dodge High Scores</a>
 		<br>
 	</div>
 </body>
