+++ views/home.ejs
@@ -14,17 +14,17 @@
 	<div class="container-fluid">
 		<br>
 		<%if (!is_mobile) { %>
-		<a href="/rocket_dodge">Rocket Dodge</a>
+		<a href="/games/rocket_dodge">Rocket Dodge</a>
 		<br>
 		<br>
 		<% } %>
-		<a href="/hold_dodge">Hold Dodge</a>
+		<a href="/games/hold_dodge">Hold Dodge</a>
 		<br>
 		<br>
-		<a href="/hold_dodge_accelerated">Hold Dodge (Accelerated)</a>
+		<a href="/games/hold_dodge_accelerated">Hold Dodge (Accelerated)</a>
 		<br>
 		<br>
-		<a href="/touch_dodge">Touch Dodge</a>
+		<a href="/games/touch_dodge">Touch Dodge</a>
 		<br>
 	</div>
 </body>
