+++ views/login.ejs
@@ -13,7 +13,7 @@
 	<br>
 	<div class="login-form text-center">
 		<h1>Login Form</h1>
-		<form action="auth" method="POST">
+		<form action="/account/auth" method="POST">
 			<input type="text" name="omniamorsaequat_username" placeholder="Username" required>
 			<input type="password" name="omniamorsaequat_password" placeholder="Password" required>
 			<input type="checkbox" id="rememberMe" name="rememberMe" title="Don't check this on public computers" checked>
@@ -21,7 +21,7 @@
 			<button type="submit">Login</button>
 		</form>
 		<br>
-		<form action="register" method="GET"><button type="submit">Register</button></form>
+		<form action="/account/register" method="GET"><button type="submit">Register</button></form>
 	</div>
 </body>
