+++ views/public_profile.ejs
@@ -25,7 +25,7 @@
                         </thead>
                         <% for (var i = 0; i < score_table.length; i++) { %>
                         <tr>
-                                <td scope="row"><a href = "/high_scores/<%= score_table[i][0] %>"><%= score_table[i][0] %></a></td>
+                                <td scope="row"><a href = "/score/high_scores/<%= score_table[i][0] %>"><%= score_table[i][0] %></a></td>
                                 <% for (var j = 1; j < score_table.length; j++) { %>
                                 <td><%= score_table[i][j] %></td>
                                 <% } %>
