+++ views/register.ejs
@@ -13,7 +13,7 @@
 	<div class="login-form text-center">
 		<h1>Register Form</h1>
 		<p>Register for high scores, scoreboards and much more!</p>
-		<form action="signup" method="POST">
+		<form action="/account/signup" method="POST">
 			<input type="text" name="username" placeholder="Username" autocomplete="off" required>
 			<input type="password" name="password" placeholder="Password" autocomplete="off" required>
 			<button type="submit">Register</button>
